Compute the accumulating transform in double precision
The rotation block of the transform that accumulates the trajectory set its
diagonal with cosf, the single precision cosine, while setting its off diagonal
with sin, the double precision sine. The result was not a rotation matrix: its
determinant was 1.71e-08 away from unity.

This runs once per scan and the result is composed into the pose estimate, so
the error compounds along the whole trajectory. Over 200 composed steps the
recovered angle drifts by 1.74e-08 with no input error whatsoever.

Fixes the first three characterisation cases.

+++ fsm_lo/include/fsm_core.hpp
@@ -509,10 +509,10 @@ class Utils
     /* Rotation matrix */
     Eigen::Matrix3d R;
     R = Eigen::Matrix3d::Identity();
-    R(0,0) = +cosf(dt);
+    R(0,0) = +cos(dt);
     R(0,1) = -sin(dt);
     R(1,0) = +sin(dt);
-    R(1,1) = +cosf(dt);
+    R(1,1) = +cos(dt);
 
     /* Compute the new transform matrix */
     Eigen::Matrix3d M_;
